Build llama.cpp within flutter build process
Rather than using prebuilt libraries, build the llama.cpp git submodule
during the regular app build process.

The library will now be installed in a standard location, which simplifies
the logic needed to load it at runtime; there is no need to ship it as an
asset.

This works on Android, and also enables the app to build and run on Linux.
Windows build is untested.

One unfortunate side effect is that when building the app in Flutter's
debug mode, the llama lib is built unoptimized and it works very very
slowly, to the point where you might suspect the app is broken.
However release mode seems as fast as before.