Update llama.cpp and move core processing to native code
Update llama.cpp to the latest version as part of an effort to make this
app usable on my Samsung Galaxy S10 smartphone.

The newer llama.cpp includes a double-close fix which was causing the app
to immediately crash upon starting the AI conversation (llama.cpp commit
47f61aaa5f76d04).

It also adds support for 3B models, which are considerably smaller. The
llama-7B models were causing Android's low memory killer to terminate
Sherpa after just a few words of conversation, whereas new models such as
orca-mini-3b.ggmlv3.q4_0.bin work on this device without quickly exhausting
all available memory.

llama.cpp's model compatibility has changed within this update, so ggml
files that were working in the previous version are unlikely to work now;
they need converting. However the orca-mini offering is already in the
new format and works out of the box.

llama.cpp's API has changed in this update. Rather than rework the Dart
code, I opted to leave it in C++, using llama.cpp's example code as a base.
Since lots of data is passed around in large arrays, it should run
considerably faster when left as native code.

This eliminates the need for Sherpa's Dart code to call llama.cpp directly,
so there's no need to separately maintain a modified version of llama.cpp
and we can use the official upstream.

Unfortunatly I ran into a snag right as I completed this work: a Dart callback
function is called from the native code when the loop is ready to receive
the next line of input, however that Dart callback function cannot be async
as far as I can see, nor is there a way to iterate the Dart main event loop
programatically, so I'm left with no way to have the processing isolate
wait on the next available user input. However the initial message is indeed
processed (as part of the initial prompt) and responded to. I'm checking this
in for reference anyway, before I rework the approach in the next commit.