Rework flow of Dart to native code
The callback from native code to Dart approach in the last commit didn't
work out, because the callback function needs to await on the next line
of user input, and we can't await from that context.

Rework the native code to return control back to Dart as soon as user
input is needed. Under this context (the isolate) we can use await.
Then call back into C when we have more input to process.

This is done by having one native code function that deals with the start
of the conversation including the prompt, and then another function that
is tasked with continuing it. This implementation duplicates a lot of code
over those two functions; I took this simplistic approach because I'm low
on time to really study and understand how this code (which originates
from llama.cpp example code) works.

Code duplication aside, the app now seems to function very nicely.